\name{bondOrderID}
\alias{bondOrderID}
\title{Conventional bond order ID number}
\description{This method calculates the conventional bond order ID number.}
\usage{
bondOrderID(g)
}
\arguments{
  \item{g}{a graph as a graphNEL object.  Each edge must have a "weight"
    data attribute containing one of the values 1 (single bond), 2
    (double bond), 3 (triple bond) or 1.5 (aromatic bond).}
}

\value{The resulting floating point number is a weighted path sum which
  takes the different bonds into account.
}
\references{% TODO Randi'c and Jurs, 1989
}
\author{Michael Schutte}

\examples{
set.seed(987)
g <- randomEGraph(LETTERS[1:10], 0.3)

edgeDataDefaults(g, "weight") <- 1
edgeData(g, "B", "I", "weight") <- 2
edgeData(g, "A", "F", "weight") <- 1.5

bondOrderID(g)
}
\keyword{ descriptors }
