\name{chemicalElements}
\docType{data}
\alias{chemicalElements}
\title{Information about chemical elements}
\description{This data set contains basic information about the chemical
elements as a data frame.}
\usage{chemicalElements}
\format{A data frame containing 117 rows representing the chemical elements
from H to Uuo.  The columns are:
\itemize{
  \item "symbol": chemical symbol.
  \item "number": atomic number.
  \item "mass": atomic mass (in unified atomic mass units).
}
}
\source{http://akiscode.com/pt/}
\keyword{ datasets }
