\name{extendedAdjacencyMatrix}
\alias{extendedAdjacencyMatrix}
\title{Extended Adjacency Matrix
}
\description{
}
\usage{
extendedAdjacencyMatrix(g)
}
\arguments{
  \item{g}{
A graph as a graphNEL object.
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}
\seealso{
}
\examples{
}
\keyword{ graphs }

