\name{polynomial}
\alias{polynomial}
\title{Polynomial Index
}
\description{
}
\usage{
polynomial(g, matrix_function, s=1)
}
\arguments{
  \item{g}{
A graph as a graphNEL object.
}
  \item{matrix_function}{
The function to calculate the desired matrix for the graph.
}
  \item{s}{
TODO
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}
\seealso{
}
\examples{
}
\keyword{ graphs }

