\name{weightStrucFuncMatrix_exp}
\alias{weightStrucFuncMatrix_exp}
\title{Weighted Structure Function Matrix
}
\description{
}
\usage{
weightStrucFuncMatrix_exp(g)
}
\arguments{
  \item{g}{
A graph as a graphNEL object.
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}
\seealso{
}
\examples{
}
\keyword{ graphs }

