\name{degreeDistribution}
\alias{degreeDistribution}
\title{Degree Distribution
}
\description{This methods calculates the degree distribution of a given graph.
}
\usage{
degreeDistribution(g, deg = NULL)
}
\arguments{
  \item{g}{a graph as a graphNEL object.
}
  \item{deg}{the distance matrix of the graph. If the parameter is empty the distance matrix will be calculated within the function.
}
}
\details{
}
\value{This methods returns the degree distribution
}
\references{

}
\author{
Laurin Mueller
}
\note{
}


\seealso{
}
\examples{

}
\keyword{graphs}
